Merge scenarios better (#25)
Goodness me, this was a rough one for such a small amount of code in this PR...

I was noticing some problems with scenario support in an isolated repository using the latest version of the library, so I decided to look into it...

It turned out we were comparing regexes with `regexA === regexB` rather than doing `regexA.toString() === regexB.toString()` and it was giving some bizarre behaviour in certain cases.

We then realised that `@types/fetch-mock` was out of date which was causing some pain.

We've also turned on overwriting for fetch routes so that the library doesn't error out too hard when you define duplicate mocks. This will only be temporary, but necessary for now. 

`findMockIndexInScenarioByMatcher()` has been removed in favour of a more functional solution.